Refactor Kalman EM workflow and workspaces
Add a SufficientStatistics type and extend KalmanWorkspace with a pred_icov cache. Rework the Kalman EM implementation: introduce _fit_kalman!, estep!, format_kf_inputs, initialize_SufficientStatistics, and split smoothing into smooth_cov!/forwards_cov!/backwards_cov! and smooth_mean!/forwards_mean!/backwards_mean!. Update precompute_kalman_constants! to refresh PDMat caches and per-trial inputs. Populate sufficient_statistics! from smoothed results. Preserve the previous long implementation by moving it to src/models/lds/kalman_old.jl for reference. Overall this change modularizes the Kalman/RTS passes, improves workspace caching, and prepares for clearer sufficient-statistics-based M-steps.
